Add support for supporting blocks
Additional rules that permit avoiding cave-ins while digging up or into
the side of dirt hills. Configurable.

Notes: There was no way to stop the dirt above my mountainside room from
caving in as soon as i poke a hole through my stone ceiling. This at
least lets me dig out a row safely while I replace the ceiling material.
It's a bit exploitable, as the additional rules (setting it to 2) will
allow the creation of "inverted dirt slopes" but natural dirt is a bit
sticky anyway, so it should be fine :)